% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.measurement.error.r
\name{anova.measurement.error}
\alias{anova.measurement.error}
\title{ANOVA for lm.rrpp model fits used in measurement.error}
\usage{
\method{anova}{measurement.error}(object, ...)
}
\arguments{
\item{object}{Object from \code{\link{lm.rrpp}}}

\item{...}{Additional lm.rrpp model fits or other arguments passed to anova.}
}
\description{
Computes an analysis of variance (ANOVA) table using 
distributions of random statistics from \code{\link{lm.rrpp}}.  
This function is the same as \code{\link{anova.lm.rrpp}} but includes statistics
specific to \code{\link{measurement.error}}, and with restrictions on
how P-values and effect.sizes are calculated.
}
\examples{
# See \code{\link{measurement.error}} examples for use.

}
\author{
Michael Collyer
}
\keyword{utilities}
