% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.r
\name{box.cox}
\alias{box.cox}
\title{Box-Cox Power Transformation}
\usage{
box.cox(y, eps = 0.001, interp = FALSE)
}
\arguments{
\item{y}{A vector of values to use.}

\item{eps}{Tolerance for lambda = 0.}

\item{interp}{Logical value for whether to spline interpolate lambda 
(ideal for vectors of small length).  This will default if the number
of values is less than 100, as in \link[MASS]{boxcox}.}
}
\value{
A list containing the following:
\item{opt.lambda}{The optimized lambda parameter.}
\item{transformed}{The power-transformed values.}
\item{lambda}{If spline interpolation is used, the values of lambda used.}
\item{loglik}{If spline interpolation is used, the log-liklihoods calculated for lambda.}
}
\description{
A function that transforms a vector in data to be 
approximately normally distributed.
}
\details{
This function uses a Box-Cox power transformation for a vector of values,
(presumably statistics from RRPP) with the following conditions:

(1) Values are first transformed as z - min(z) + 1, where z 
is a vector of standard deviates of the original values.  
This assures values are positive
with a minimum value of 1, such that a lambda of 0 returns a 
minimum transformed value of 0.  This also allows transformation of
negative values.


(2) The lambda parameter is optimized over a range of -2 to +2,
as is typical for most applications.

(3) The first value (assumed to be the observed value in a 
distribution of random values) is removed from the vector to 
perform optimization of lambda, but is returned to the vector for
transformation.  This assures that large, aberrant values, such
as might be found via RRPP for large effects, do not have 
leverage for the optimization.

This function is similar to \link[MASS]{boxcox} in most regards
but uses a consistent initial scaling and shifting of values for comparative
purposes, when calculating effect sizes with \link{effect.size}.  In essence,
it is the same as using the \link[MASS]{boxcox} function, after
first scaling values and adding 1.  However, it also assumes
the first value is an observed value in a vector of random statistics.  It
is intended for RRPP functions, only, but could be used with other results.  
However, one must be aware that the first value will be removed for optimization.
}
\references{
Box, G. E. P. and Cox, D. R. (1964) An analysis of transformations 
(with discussion). Journal of the Royal Statistical Society B, 26, 211–252.
}
\author{
Michael Collyer
}
\keyword{utilities}
