% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRPP.support.code.r
\name{center}
\alias{center}
\title{Support function for RRPP}
\usage{
center(x)
}
\arguments{
\item{x}{Matrix to mean-center.}
}
\description{
Center a matrix, quickly
}
\author{
Michael Collyer
}
\keyword{utilities}
