% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.defunct.R
\name{classify}
\alias{classify}
\title{Deprecated functions in RRPP}
\usage{
classify()
}
\description{
The following function has been deprecated in RRPP
}
\details{
This function has been deprecated. Use \code{\link{prep.lda}} instead.
}
