% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.lm.rrpp.r
\name{coef.lmm.rrpp}
\alias{coef.lmm.rrpp}
\title{coef for lmm.rrpp model fits}
\usage{
\method{coef}{lmm.rrpp}(object, type = c("matrix", "list"), ...)
}
\arguments{
\item{object}{Object from \code{\link{lmm.rrpp}}}

\item{type}{Whether to combine coefficients in a matrix or a list by variable.}

\item{...}{Other arguments (currently none)}
}
\description{
The coef.lmm.rrpp function computes 
combined fixed and random effects by subject for 
\code{\link{lmm.rrpp}} model fits. This function does not operate the same as
\code{\link{coef.lm.rrpp}}, for which coefficients and fixed effects are the same,
and tests can be performed on fixed effects.  It is merely a wrapper to combine 
the results of \code{\link{ranef.lmm.rrpp}} and \code{\link{fixef.lmm.rrpp}}.
}
\author{
Michael Collyer
}
\keyword{utilities}
