% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.r
\name{effect.size}
\alias{effect.size}
\title{Obtain Effect-size from a vector of values}
\usage{
effect.size(x, center = TRUE)
}
\arguments{
\item{x}{The vector of data to use.}

\item{center}{Logical value for whether to center x.}

\item{...}{Arguments passed to \link{powerTrans}.}
}
\description{
A function to find the effect size (Z-score) of a target,
from a vector of values presumably obtained in random permutations.
}
\author{
Michael Collyer
}
\keyword{utilities}
