% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRPP.support.code.r
\name{fast.solve}
\alias{fast.solve}
\title{Support function for RRPP}
\usage{
fast.solve(x)
}
\arguments{
\item{x}{Matrix to invert.}
}
\description{
Invert a matrix, quickly.  Functions similar to solve function, but much quicker.
}
\author{
Michael Collyer
}
\keyword{utilities}
