% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRPP.data.r
\docType{data}
\name{fishy}
\alias{fishy}
\title{Simulated Fish Data for Measurement Error Analysis}
\description{
Simulated Fish Data for Measurement Error Analysis
}
\details{
Data as simulated in Collyer and Adams (2024),
resembling a fish shape, comprising Procrustes coordinates
for 11 anatomical landmarks.  Data represent 120 
configurations for 60 subjects, each with two replicates of
measurement.  The 60 subjects represent 20 subjects each 
from three groups.
}
\references{
Collyer, M.L, and D.C. Adams. 2024. 
Interrogating random and systematic measurement 
error in morphometric data. Evolutionary Biology. 
In press.
}
\author{
Michael Collyer
}
\keyword{datasets}
