% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRPP.utils.r
\name{fixed.effects.lmm.rrpp}
\alias{fixed.effects.lmm.rrpp}
\title{fixed.effects for lmm.rrpp model fits}
\usage{
fixed.effects.lmm.rrpp(object, ...)
}
\arguments{
\item{object}{Object from \code{\link{lmm.rrpp}}}

\item{...}{Other arguments passed.}
}
\description{
Returns fixed effects from a lmm.rrpp object.  Alias for fixef.lmm.rrpp.
}
\author{
Michael Collyer
}
\keyword{utilities}
