% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRPP.utils.r
\name{fixef.lmm.rrpp}
\alias{fixef.lmm.rrpp}
\title{fixef for lmm.rrpp model fits}
\usage{
fixef.lmm.rrpp(object, ...)
}
\arguments{
\item{object}{Object from \code{\link{lmm.rrpp}}}

\item{...}{Other arguments passed.}
}
\description{
Returns fixed effects from a lmm.rrpp object.
}
\author{
Michael Collyer
}
\keyword{utilities}
