% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRPP.data.r
\docType{data}
\name{flower}
\alias{flower}
\title{Flower Production and Temperature}
\description{
A data set containing variables for flower production and
temperature, averaged yearly for 30 years, from a dry tropical forest.
}
\details{
The data in this data set are samples from a larger study 
that compared flower production between forest types over 30 years, 
based on climate variables.  The data in this set are from a dry tropical forest,
include variables for annual flower production, temperature, and year. These 
data are suitable for considering the temporal non-independence of observations in 
generalized least squares (GLS) estimation, for the example in the book,
Permutational Biometry (Collyer & Adams, 2025).
}
\references{
Pau, S., Wolkovich, E. M., Cook, B. I., Nytch, C. J., Regetz, J., Zimmerman,
J. K., and Joseph Wright, S. (2013). Clouds and temperature drive dynamic
changes in tropical flower production. Nature Climate Change, 3:838–842.

Collyer, M. L., & Adams D. C. (2025). Permutational Biometry: Volume
1, Univariate Data. Iowa State University Digital Press.  NOT YET PUBLISHED.
}
\keyword{datasets}
