% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRPP.utils.r
\name{focusMEonSubject}
\alias{focusMEonSubject}
\title{Plot Function for RRPP}
\usage{
focusMEonSubject(x, subject = NULL, ...)
}
\arguments{
\item{x}{Plot from \code{\link{plot.measurement.error}}}

\item{subject}{The specific subject to plot}

\item{...}{Other arguments passed onto plot}
}
\description{
Reduces a plot.measurement.error to a single research subject.  This can be 
for cases when many overlapping subjects in a plot obscure interpretation 
for specific subjects.
}
\author{
Michael Collyer
}
\keyword{utilities}
