% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRPP.utils.r
\name{focusMEonSubjects}
\alias{focusMEonSubjects}
\title{Plot Function for RRPP}
\usage{
focusMEonSubjects(x, subjects = NULL, shadow = TRUE, ...)
}
\arguments{
\item{x}{Plot from \code{\link{plot.measurement.error}}}

\item{subjects}{The specific subject to plot}

\item{shadow}{A logical value for whether to show other subject values as
shadows of their locations.}

\item{...}{Other arguments passed onto plot}
}
\description{
Reduces a plot.measurement.error to a single research subject.  This can be 
for cases when many overlapping subjects in a plot obscure interpretation 
for specific subjects.
}
\author{
Michael Collyer
}
\keyword{utilities}
