% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRPP.utils.r
\name{getModelCov}
\alias{getModelCov}
\title{Utility Function for RRPP}
\usage{
getModelCov(fit, type = c("Cov", "Pcov"), format = c("S3", "S4"))
}
\arguments{
\item{fit}{Object from \code{\link{lm.rrpp}}}

\item{type}{Whether the Cov or Pcov matrix is returned}

\item{format}{Whether an S3 or S4 format is returned}
}
\description{
A function mostly for internal processing but can be used to extract either
the covariance matrix (Cov) or the projection matrix for transformations made 
from the covariance matrix (Pcov), which is basically the square-root of 
the covariance matrix.  There is also an option for S3 or S4 versions.
}
\author{
Michael Collyer
}
\keyword{utilities}
