% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRPP.utils.r
\name{getPermInfo}
\alias{getPermInfo}
\title{Utility Function for RRPP}
\usage{
getPermInfo(
  fit,
  attribute = c("perms", "perm.method", "block", "perm.seed", "perm.schedule", "all")
)
}
\arguments{
\item{fit}{Object from \code{\link{lm.rrpp}}}

\item{attribute}{The various attributes that are used to generate RRPP
permutation schedules.  If there are n observations, each iteration has
some randomization of 1:n, restricted by the arguments that match attributes
provided by this function.}
}
\description{
A function mostly for internal processing but can be used to extract
RRPP permutation information for other reasons.
}
\author{
Michael Collyer
}
\keyword{utilities}
