% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRPP.utils.r
\name{getResCov}
\alias{getResCov}
\title{Utility Function for RRPP}
\usage{
getResCov(fit, useDf = TRUE, standardize = FALSE)
}
\arguments{
\item{fit}{Object from \code{\link{lm.rrpp}}}

\item{useDf}{Logical value for whether the degrees of freedom from the linear model
fit should be used (if TRUE), as opposed to the number of observations (if FALSE).}

\item{standardize}{Logical value for whether residuals should be standarized.  If TRUE,
a correlation matrix is produced.}
}
\description{
A function mostly for internal processing but can be used to extract the residual
covariance matrix.  This matrix is the residual covariance matrix,
not the model covariance matrix used for estimation (see \code{\link{getModelCov}}). 
Options for averaging over degrees of freedom or number of
observations, plus standardization, are also available.
}
\author{
Michael Collyer
}
\keyword{utilities}
