% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRPP.utils.r
\name{getTerms}
\alias{getTerms}
\title{Utility Function for RRPP}
\usage{
getTerms(fit)
}
\arguments{
\item{fit}{Object from \code{\link{lm.rrpp}}}
}
\description{
A function mostly for internal processing but can be used to extract
The terms for each reduced and full model used in an \code{\link{lm.rrpp}} fit.
}
\author{
Michael Collyer
}
\keyword{utilities}
