% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.rrpp.r
\name{lm.rrpp}
\alias{lm.rrpp}
\title{Linear Model Evaluation with a Randomized Residual Permutation Procedure}
\usage{
lm.rrpp(f1, iter = 999, seed = NULL, int.first = FALSE, RRPP = TRUE,
  SS.type = c("I", "II", "III"), data = NULL, Cov = NULL,
  Cov.par = NULL, print.progress = TRUE, Parallel = FALSE, ...)
}
\arguments{
\item{f1}{A formula for the linear model (e.g., y~x1+x2).  Can also be a linear model fit
from \code{\link{lm}}.}

\item{iter}{Number of iterations for significance testing}

\item{seed}{An optional argument for setting the seed for random permutations of the resampling procedure.
If left NULL (the default), the exact same P-values will be found for repeated runs of the analysis (with the same number of iterations).
If seed = "random", a random seed will be used, and P-values will vary.  One can also specify an integer for specific seed values,
which might be of interest for advanced users.}

\item{int.first}{A logical value to indicate if interactions of first main effects should precede subsequent main effects}

\item{RRPP}{A logical value indicating whether residual randomization should be used for significance testing}

\item{SS.type}{A choice between type I (sequential), type II (hierarchical), or type III (marginal)
sums of squares and cross-products computations.}

\item{data}{A data frame for the function environment, see \code{\link{rrpp.data.frame}}}

\item{Cov}{An optional argument for including a covariance matrix to address the non-independence
of error in the estimation of coefficients (via GLS).}

\item{Cov.par}{An optional argument to override the number of parameters required for an optional
covariance matrix argument.  This argument would be used if the correlation structure of the covariance
matrix is not obvious but should be structured.  For example, if an autoregressive correlation structure
is used to estimate the covariance matrix with the \code{\link{corClasses}} set of functions, 
and the number of parameters should be 1, it is not possible to detect this with a p x p matrix for p variables.
(It is also not feasible to generate covariance matrices outside of the \code{nlme} package.  Users would have to do this
prior to using \code{lm.rrpp}.)  In an example like this, one should set
 the number of parameters to 1.  Otherwise, the number of parameters will be p(p+1)/2,
assuming an unstructured covariance matrix.  This argument will have no influence on the estimation of coefficients
or SS, but could have impact for downstream model comparisons, using \code{\link{compare.models}}.}

\item{print.progress}{A logical value to indicate whether a progress bar should be printed to the screen.
This is helpful for long-running analyses.}

\item{Parallel}{A logical value to indicate whether parallel processing should be used.  If TRUE, this argument
invokes forking of processor cores, using the \code{parallel} library.  This option is only available to unix systems
and should only be used for rather long analyses (that would normally take over 10 seconds on a single core).  Currently,
parallel processing is performed on all but one core with no option to change the number of cores.  Systems with Windows
platforms will automatically dafualt to a single-core application of this function.}

\item{...}{Arguments typically used in \code{\link{lm}}, such as weights or offset, passed on to
\code{rrpp.fit} for estimation of coefficients.}
}
\value{
An object of class \code{lm.rrpp} is a list containing the following
\item{call}{The matched call.}
\item{LM}{Linear Model objects, including data (Y), coefficients, design matrix (X), sample size
(n), number of dependent variables (p), QR decomposition of the design matrix, fitted values, residuals,
weights, offset, model terms, model frame (data), random coefficients (through permutations),
random vector distances for coefficients (through permutations), and whether OLS or GLS was performed}
\item{ANOVA}{Analysis of variance objects, including the SS type, random SS outcomes, random MS outcomes,
random R-squared outcomes, random F outcomes, random Cohen's f-squared outcomes, P-values based on random F
outcomes, effect sizes for random outcomes, sample size (n), number of variables (p), and degrees of freedom for
model terms (df).  These objects are used to construct ANOVA tables.}
\item{PermInfo}{Permutation procedure information, including the number of permutations (perms), The method
of resdiual randomization (perm.method), and each permutation's sampling frame (perm.schedule), which
is a list of reordered sequences of 1:n, for how residuals were randomized.}
}
\description{
Function performs a linear model fit over many random permutations of data, using
a randomized residual permutation procedure.
}
\details{
The function fits a linear model using ordinary least squares (OLS) or generalized
least squares (GLS) estimation of coefficients over any number of random permutations of
the data.  A permutation procedure that randomizes vectors of residuals is emplyed.  This
procedure can randomize two types of resdiuals: residuals from null models or residuals from
an intercept model.  The latter is the same as randomizing full values, and is referred to as
as a full randomization permutation procedure (FRPP); the former uses the residuals from null
models, which are defined by the type of sums of squares and cross-products (SSCP) sought in an
analysis of variance (ANOVA), and is referred to as a randomized residual permutation procedure (RRPP).
Types I, II, and III SSCPs are supported.

Users define the SSCP type, the permutation proecure type, whether a covariance matrix is included
(GLS estimation), and a few arguments related to computations.  Analytical results comprise observed linear model
results (coefficients, fitted values, residuals, etc.), random sums of squares (SS) across permutation iterations,
and other parameters for perfoming ANOVA and other hypothesis tests, using
empirically-derived probability distributions.

\code{lm.rrpp} emphasizes estimation of effect sizes with standard deviates of observed statistics
from distributions of random outcomes.  When performing ANOVA, using the \code{\link{anova}} function,
the effect type (statistic choice) can be varied.  See \code{\link{anova.lm.rrpp}} for more details.  Please
recognize that the type of SS must be chosen prior to running \code{lm.rrpp} and not when applying \code{\link{anova}}
to the \code{lm.rrpp} fit, as design matrices for the linear model must be created first.  Therefore, SS.type
is an argument for \code{lm.rrpp} and effect.type is an argument for \code{\link{anova.lm.rrpp}}.
}
\examples{
### Example of OLS/GLS plus SS.type options
# Create a phylogenetic covariance matrix plus data from a random phylogeny,
# and perform OLS and GLS analyses
# UPDATE EXAMPLE
}
\references{
Anderson MJ. 2001. A new method for non-parametric multivariate analysis of variance.
   Austral Ecology 26: 32-46.

Anderson MJ. and C.J.F. terBraak. 2003. Permutation tests for multi-factorial analysis of variance.
   Journal of Statistical Computation and Simulation 73: 85-113.

Collyer, M.L., D.J. Sekora, and D.C. Adams. 2015. A method for analysis of phenotypic change for phenotypes described
by high-dimensional data. Heredity. 115:357-365.

Adams, D.C. and M.L. Collyer. 2016.  On the comparison of the strength of morphological integration across morphometric
datasets. Evolution. 70:2623-2631.
}
\seealso{
\code{procD.lm} and \code{procD.pgls} within \code{geomorph}; \code{\link[stats]{lm}} for more on linear model fits.
}
\author{
Michael Collyer
}
\keyword{analysis}
