% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/looCV.r
\name{looCV_table}
\alias{looCV_table}
\title{Table for looCV comparisons}
\usage{
looCV_table(...)
}
\arguments{
\item{...}{Any number of \code{\link{looCv}} objects}
}
\description{
Function simply produces a table of MSE and cross-validated
MSE for any number of compared models.  The table is sorted
from lowest to highest cross-validated MSE.  The looCV objects
must use method = 'fit' or a table cannot be generated.
}
\examples{

# Example with snakeHS data

data(snakeHS)
fitS <- lm.rrpp(HS ~ SVL, data = snakeHS,
                iter = 0)
fitSR <- lm.rrpp(HS ~ SVL + Region, data = snakeHS,
                 iter = 0)
fitSxR <- lm.rrpp(HS ~ SVL * Region, data = snakeHS,
                  iter = 0)

CVS <- looCV(fitS)
CVSR <- looCV(fitSR)
CVSxR <- looCV(fitSxR)

summary(CVS)
summary(CVSR)
summary(CVSxR)

looCV_table(CVS, CVSR, CVSxR)

}
\seealso{
\code{\link{looCV}}, \code{\link{model.comparisons}}
}
\author{
Michael Collyer
}
\keyword{analysis}
