% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRPP.data.r
\docType{data}
\name{mammalHR}
\alias{mammalHR}
\title{Mammal Home Range Data}
\description{
Home range and body mass data for 49 species of mammals.
}
\details{
The data set is from a study of mammal species home range size 
(squared kilometers) and its covariation with body mass or whether mammals
are carnivores or ungulates (Garland et al., 1993).  the data set also includes a
covariance matrix based on phylogenetic relatedness among species, to account for the
non-independence of observations.
These data are suitable data for demonstrating generalized least squares (GLS)
estimation, used in the book, 
Permutational Biometry (Collyer & Adams, 2025).
}
\references{
Garland, T., Dickerman, A. W., Janis, C. M., and Jones, J. A. (1993). 
Phylogenetic analysis of covariance by computer simulation. Systematic Biology,
42:265–292.

Collyer, M. L., & Adams D. C. (2025). Permutational Biometry: Volume
1, Univariate Data. Iowa State University Digital Press.  NOT YET PUBLISHED.
}
\keyword{datasets}
