% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRPP.data.r
\docType{data}
\name{mammalHRphy}
\alias{mammalHRphy}
\title{Phylogenetic Data for the Mammal Home Range Data}
\description{
Phylogenetic tree data that can be used along with 
\code{\link{mammalHR}}.
}
\details{
A class \code{phylo} object containing data for the phylogenetic tree
that pertains to the \code{\link{mammalHR}} data set.  Although these data are
not required for the generalized least squares (GLS)
estimation example, used in the book, 
Permutational Biometry (Collyer & Adams, 2025), they provide additional 
graphical support for the example.
}
\references{
Collyer, M. L., & Adams D. C. (2025). Permutational Biometry: Volume
1, Univariate Data. Iowa State University Digital Press.  NOT YET PUBLISHED.
}
\keyword{datasets}
