% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRPP.utils.r
\name{model.frame.lm.rrpp}
\alias{model.frame.lm.rrpp}
\title{Extract model frame from a lm.rrpp object}
\usage{
\method{model.frame}{lm.rrpp}(formula, ...)
}
\arguments{
\item{formula}{Object from \code{\link{lm.rrpp}}}

\item{...}{further arguments passed to or from other methods}
}
\description{
\code{model.frame.lm.rrpp} returns the model frame constructed for an \code{lm.rrpp} object.
}
\author{
Michael Collyer
}
\keyword{utilities}
