% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinate.r
\name{ordinate}
\alias{ordinate}
\title{Ordination tool for data aligned to another matrix}
\usage{
ordinate(Y, A = NULL, Cov = NULL, scale. = FALSE, tol = NULL,
  rank. = NULL, newdata = NULL)
}
\arguments{
\item{Y}{An n x p data matrix.}

\item{A}{An optional n x n symmetric matrix or an n x k data matrix, where k is the number of variables that could
be associated with the p variables of Y.  If NULL, an n x n identity matrix will be used.}

\item{scale.}{a logical value indicating whether the variables should be scaled to have unit variance before the analysis 
takes place. The default is FALSE.}

\item{tol}{A value indicating the magnitude below which components should be omitted. (Components are omitted if their 
standard deviations are less than or equal to tol times the standard deviation of the first component.) 
With the default null setting, no components are omitted (unless rank. is specified less than min(dim(x)).). 
Other settings for tol could be tol = 0 or tol = sqrt(.Machine$double.eps), which would omit essentially constant components.
This argument is exactly the same as in \code{\link{prcomp}}}

\item{rank.}{Aptionally, a number specifying the maximal rank, i.e., maximal number of aligned components to be used. 
This argument can be set as alternative or in addition to tol, useful notably when the desired rank is considerably 
smaller than the dimensions of the matrix.  This argument is exactly the same as in \code{\link{prcomp}}}

\item{newdata}{An optional data frame of values for the same variables of Y to be projected onto 
aligned components.}

\item{A}{An optional n x n covariance matrix to describe the non-independence among
observations in Y, and provide a GLS-centering of data.  Note that Cov and A can be the same, if one
wishes to align GLS residuals to the same matrix used to obtain them.}
}
\value{
An object of class \code{ordinate} is a list containing the following
\item{x}{Aigned component scores for all observations}
\item{xn}{Optional projection of new data onto components.}
\item{d}{The portion of the singular values attributed to the aligned components.}
\item{sdev}{Standard deviations of d; i.e., the scale of the components.}
\item{rot}{The matrix of variable loadings, i.e. the singular vectors, \bold{V}.}
\item{center}{The OLS or GLS means vector used for centering.}
\item{scale}{The sclaing used, or FALSE.}
\item{alignment}{Whether data were aligned to principal axes or the name of another matrix.}
\item{GLS}{A logical value to indicate if GLS-centering and projection was used.}
}
\description{
Function performs a singular value decomposition of ordinary least squares (OLS) or 
generalized least squares (GLS) residuals, aligned to an alternative matrix, plus projection of 
data onto vectors obtained.
}
\details{
The function performs a singular value decomosition, \bold{A'Z} = \bold{UDV'}, where
\bold{Z} is a matrix of residuals (obtained from \bold{Y} - see below) and \bold{A}
is an alignment matrix with the same number of rows as \bold{Z}.  
(\bold{'} indicates matrix transposition.)  \bold{U}  and \bold{V}
are the matrices of left and right singular vectors, and \bold{D} is a diagonal matrix of singular
values. \bold{V} are the vectors that describe maximized covariation between \bold{Y} and \bold{A}. 
If \bold{A} = \bold{I}, an n x n identity matrix, \bold{V} are the
eigen vectors (principal components) of \bold{Y}.

\bold{Z} represents a centered and potentially standardized form of \bold{Y}.  This
function can center data via OLS or GLS means (the latter if a covariance matrix  to describe
the non-independence among observations is provided).  If standardizing variables is preferred,
then \bold{Z} both centers and scales the vectors of \bold{Y} by their standard deviations.

Data are projected onto aligned vectors, \bold{ZV}, which in the case of OLS residuals is
an orthogonal projection and in the case of GLS is an oblique projection.

The versatility of using an alignment approach is that alternive data space rotations are possible.
Principal components are thus the vectors that maximize variance with respect to the data, themselves,
but "components" of (co)variation can be described for any inter-matrix relationship, including
phylogenetic signal, ecological signal, ontogenetic signal, size allometry, etc.
More details are provided in Collyer and Adams (in review).

Much of this function is consisten with the \code{\link{prcomp}} function, except that centering data
is not an option (it is required).
}
\examples{

# Examples use residuals from a regression of salamander morphological 
# traits against body size (snout to vent length, SVL).
# Observations are species means and a phylogenetic covariance matrix
# describes the relatedness among observations.

data("PlethMorph")
R <- lm.rrpp(cbind (TailLength, HeadLength, Snout.eye, BodyWidth, 
Forelimb, Hindlimb) ~ SVL, iter = 0, 
data = PlethMorph, print.progress = FALSE)$LM$residuals

PCA.ols <- ordinate(R, scale. = TRUE)
PCA.ols$rot
prcomp(R, scale. = TRUE)$rotation # should be the same

PCA.gls <- ordinate(R, scale. = TRUE, Cov = PlethMorph$PhyCov)

# Align to phylogenetic signal

PaCA.ols <- ordinate(R, A = PlethMorph$PhyCov, scale. = TRUE)
PaCA.gls <- ordinate(R, A = PlethMorph$PhyCov, scale. = TRUE,
Cov = PlethMorph$PhyCov)

# Summaries

summary(PCA.ols)
summary(PCA.gls)
summary(PaCA.ols)
summary(PaCA.gls)

# Plots

par(mfrow = c(2,2))
plot(PCA.ols, main = "PCA OLS")
plot(PCA.gls, main = "PCA GLS")
plot(PaCA.ols, main = "PaCA OLS")
plot(PaCA.gls, main = "PaCA GLS")
par(mfrow = c(1,1))

}
\references{
Collyer and Adams, in prep.
}
\seealso{
\code{\link{prcomp}} and \code{gm.prcomp} within \code{geomorph}
}
\author{
Michael Collyer
}
\keyword{analysis}
