% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRPP.data.r
\docType{data}
\name{PlethCinHoff}
\alias{PlethCinHoff}
\title{Plethodon comparative SVL data}
\description{
Snout-to-vent-length (SVL) data for two species of salamanders
}
\details{
The data set is a subset of data from two species of salamanders 
(Plethodon cinereus and 
P. hoffmani) in eastern North America (based on Adams and Rohlf, 2000; 
Adams, 2000), collected from the Smithsonian Museum (Washington DC, USA).  Data 
include species, locality (whether in sympatry or allopatry), and SVL in mm.
These data are suitable data for demonstrating single factor and
factorial analysis of variance (ANOVA).
(ANCOVA), used in the book, Permutational Biometry (Collyer & Adams, 2025).
}
\references{
Adams, D. C. (2000). Divergence of trophic morphology and resource use
among populations of Plethodon cinereus and P. hoffmani in Pennsylvania.
In Bruce, R. C., Jaeger, R. G., and Houck, L. D., editors, The Biology of
Plethodontid Salamanders, pages 383–394. Springer US.

Adams, D. C. and Rohlf, F. J. (2000). Ecological character displacement in
Plethodon: Biomechanical differences found from a geometric morphometric
study. Proceedings of the National Academy of Sciences, 97:4106–4111.

Collyer, M. L., & Adams D. C. (2025). Permutational Biometry: Volume
1, Univariate Data. Iowa State University Digital Press.  NOT YET PUBLISHED.
}
\author{
Dean Adams
}
\keyword{datasets}
