% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRPP.utils.r
\name{plot.interSubVar}
\alias{plot.interSubVar}
\title{Plot Function for RRPP}
\usage{
\method{plot}{interSubVar}(x, ...)
}
\arguments{
\item{x}{Object from \code{\link{inteSubVar}}}

\item{...}{Arguments passed onto \code{\link{image}} and \code{\link{plot}}.}
}
\description{
This function produces a heat map for inter-subject variability, based on 
results from a \code{\link{measurement.error}} object.  The function, 
\code{\link{interSubVar}}, must first be used on the \code{\link{measurement.error}} 
object to obtain variability statistics.  This function use the \code{\link{image}}
function to produce plots.  It does little to manipulate such plots, but any
argument for \code{\link{image}} can be manipulated here, as well as the graphical
parameters that can be adjusted within \code{\link{image}}.
}
\author{
Michael Collyer
}
\keyword{utilities}
