% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRPP.utils.r
\name{plot.model.comparison}
\alias{plot.model.comparison}
\title{Plot Function for RRPP}
\usage{
\method{plot}{model.comparison}(x, ...)
}
\arguments{
\item{x}{plot object (from \code{\link{model.comparison}})}

\item{...}{other arguments passed to plot (helpful to employ
different colors or symbols for different groups).  See
\code{\link{plot.default}} and \code{\link{par}}}
}
\description{
Plot Function for RRPP
}
\author{
Michael Collyer
}
\keyword{utilities}
\keyword{visualization}
