% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRPP.utils.r
\name{plot.ordinate}
\alias{plot.ordinate}
\title{Plot Function for RRPP}
\usage{
\method{plot}{ordinate}(x, axis1 = 1, axis2 = 2, ...)
}
\arguments{
\item{x}{An object of class \code{\link{ordinate}}}

\item{axis1}{A value indicating which component should be displayed as the X-axis (default = C1)}

\item{axis2}{A value indicating which component should be displayed as the Y-axis (default = C2)}

\item{...}{other arguments passed to plot (helpful to employ
different colors or symbols for different groups).  See}
}
\value{
An object of class "plot.ordinate" is a list with components
 that can be used in other plot functions, such as the type of plot, points, 
 a group factor, and other information depending on the plot parameters used.
}
\description{
Plot Function for RRPP
}
\author{
Michael Collyer
}
\keyword{utilities}
\keyword{visualization}
