% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRPP.utils.r
\name{plot.predict.lm.rrpp}
\alias{plot.predict.lm.rrpp}
\title{Plot Function for RRPP}
\usage{
\method{plot}{predict.lm.rrpp}(x, PC = FALSE, ellipse = FALSE,
  label = TRUE, ...)
}
\arguments{
\item{x}{plot object (from \code{\link{lm.rrpp}})}

\item{PC}{A logical argument for whether the data space should be rotated to its 
principal components}

\item{ellipse}{A logical argument to change error bars to ellipses in multivariate plots.
It has no function for univariate plots.}

\item{label}{A logical argument for whether points should be labled 
(in multivariate plots).}

\item{...}{other arguments passed to plot (helpful to employ
different colors or symbols for different groups).  See
\code{\link{plot.default}} and \code{\link{par}}}
}
\description{
Plot Function for RRPP
}
\author{
Michael Collyer
}
\keyword{utilities}
\keyword{visualization}
