% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRPP.utils.r
\name{plot.z.model.test}
\alias{plot.z.model.test}
\title{Plot Function for RRPP}
\usage{
\method{plot}{z.model.test}(x, label = c("none", "model", "form"), ...)
}
\arguments{
\item{x}{plot object (from \code{\link{z.model.test}})}

\item{label}{One of "form", "model", or "none", to indicate where to lable
points by formulae, model sequence, or not at all.}

\item{...}{other arguments passed to plot (helpful to employ
different colors or symbols for different groups).  See
\code{\link{plot.default}} and \code{\link{par}}}
}
\description{
Plot Function for RRPP
}
\author{
Michael Collyer
}
\keyword{utilities}
\keyword{visualization}
