% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRPP.data.r
\docType{data}
\name{podarcis}
\alias{podarcis}
\title{Podarcis Wall Lizards from the Mediterranean Basin}
\description{
A data set of body size for males and females form different lizard 
species in different geographic locations.
}
\details{
These data are a subset of the data collected originally by 
Kaliontzopoulou et al.(2015).  Variables include snout-to-vent-length (SVL),
sex, species, and habitat preference.  These 
data are suitable for comparing effect sizes in the book,
Permutational Biometry (Collyer & Adams, 2025).
}
\references{
Kaliontzopoulou, A., Carretero, M. A., and Adams, D. C. (2015). Ecomorphological
variation in male and female wall lizards and the macroevolution
of sexual dimorphism in relation to habitat use. Journal of Evolutionary
Biology, 28:80–94.

Collyer, M. L., & Adams D. C. (2025). Permutational Biometry: Volume
1, Univariate Data. Iowa State University Digital Press.  NOT YET PUBLISHED.
}
\keyword{datasets}
