% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.lm.rrpp.r
\name{predict.lm.rrpp}
\alias{predict.lm.rrpp}
\title{predict for lm.rrpp model fits}
\usage{
\method{predict}{lm.rrpp}(object, newdata, confidence = 0.95, ...)
}
\arguments{
\item{object}{Object from \code{\link{lm.rrpp}}.}

\item{newdata}{Data frame of either class \code{\link{data.frame}} or \code{\link{rrpp.data.frame}}.  If null,
the data frame from the lm.rrpp fit will be used, effectively calculating all fitted values and
their confidence intervals.}

\item{confidence}{The desired confidence interval level for prediction.}

\item{...}{Other arguments (currently none)}
}
\description{
Computes predicted values from an \code{\link{lm.rrpp}} model fit, using bootstrapped residuals
to generate confidence intervals.  The bootstrap permutations use the same number of iterations and seed as used
in the \code{\link{lm.rrpp}} model fit. A \code{\link{predict.lm.rrpp}} object can be plotted using various options.
See \code{\link{plot.predict.lm.rrpp}}.
}
\examples{
# See examples for lm.rrpp
}
\author{
Michael Collyer
}
\keyword{utilities}
