% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRPP.utils.r
\name{print.anova.lm.rrpp}
\alias{print.anova.lm.rrpp}
\title{Print/Summary Function for RRPP}
\usage{
\method{print}{anova.lm.rrpp}(x, ...)
}
\arguments{
\item{x}{print/summary object (from \code{\link{lm.rrpp}})}

\item{...}{other arguments passed to print/summary}
}
\description{
Print/Summary Function for RRPP
}
\author{
Michael Collyer
}
\keyword{utilities}
