% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRPP.utils.r
\name{print.betaTest}
\alias{print.betaTest}
\title{Print/Summary Function for RRPP}
\usage{
\method{print}{betaTest}(x, ...)
}
\arguments{
\item{x}{Object from \code{\link{betaTest}}}

\item{...}{Other arguments passed onto betaTest}
}
\description{
Print/Summary Function for RRPP
}
\author{
Michael Collyer
}
\keyword{utilities}
