% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRPP.utils.r
\name{print.classify}
\alias{print.classify}
\title{Print/Summary Function for RRPP}
\usage{
\method{print}{classify}(x, ...)
}
\arguments{
\item{x}{Object from \code{\link{classify}}}

\item{...}{Other arguments passed onto classify}
}
\description{
Print/Summary Function for RRPP
}
\author{
Michael Collyer
}
\keyword{utilities}
