% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRPP.utils.r
\name{print.compare.models}
\alias{print.compare.models}
\title{compare.models}
\usage{
\method{print}{compare.models}(x, ...)
}
\arguments{
\item{x}{print/summary object (from \code{\link{lm.rrpp}})}

\item{...}{other arguments passed to print/summary}
}
\description{
Print/Summary Function for RRPP
}
\author{
Michael Collyer
}
\keyword{utilities}
