% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRPP.utils.r
\name{print.ICCstats}
\alias{print.ICCstats}
\title{Print/Summary Function for RRPP}
\usage{
\method{print}{ICCstats}(x, ...)
}
\arguments{
\item{x}{print/summary object (from \code{\link{ICCstats}})}

\item{...}{other arguments passed to print/summary}
}
\description{
Print/Summary Function for RRPP
}
\author{
Michael Collyer
}
\keyword{utilities}
