% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRPP.utils.r
\name{print.kcomp}
\alias{print.kcomp}
\title{Print/Summary Function for RRPP}
\usage{
\method{print}{kcomp}(x, ...)
}
\arguments{
\item{x}{Object from \code{\link{kcomp}}}

\item{...}{Other arguments passed onto print.kcomp}
}
\description{
Print/Summary Function for RRPP
}
\author{
Michael Collyer
}
\keyword{utilities}
