% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRPP.utils.r
\name{print.looCV}
\alias{print.looCV}
\title{Print/Summary Function for RRPP}
\usage{
\method{print}{looCV}(x, ...)
}
\arguments{
\item{x}{Object from \code{\link{looCV}}}

\item{...}{Other arguments passed onto print.looCV}
}
\description{
Print/Summary Function for RRPP
}
\author{
Michael Collyer
}
\keyword{utilities}
