% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRPP.utils.r
\name{print.lr_test}
\alias{print.lr_test}
\title{Print/Summary Function for RRPP}
\usage{
\method{print}{lr_test}(x, ...)
}
\arguments{
\item{x}{Object from \code{\link{lr_test}}}

\item{...}{Other arguments passed onto print.lr_test}
}
\description{
Print/Summary Function for RRPP
}
\author{
Michael Collyer
}
\keyword{utilities}
