% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRPP.utils.r
\name{print.pairwise.model.Z}
\alias{print.pairwise.model.Z}
\title{Print/Summary Function for RRPP}
\usage{
\method{print}{pairwise.model.Z}(x, stats.table = TRUE, ...)
}
\arguments{
\item{x}{print/summary object (from \code{\link{pairwise.model.Z}})}

\item{stats.table}{A logical value for whether to condense results into
a single table of multiple statistics.}

\item{...}{other arguments passed to print/summary}
}
\description{
Print/Summary Function for RRPP
}
\author{
Dean Adams and Michael Collyer
}
\keyword{utilities}
