% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRPP.utils.r
\name{print.pairwise}
\alias{print.pairwise}
\title{Print/Summary Function for RRPP}
\usage{
\method{print}{pairwise}(x, ...)
}
\arguments{
\item{x}{Object from \code{\link{pairwise}}}

\item{...}{Other arguments passed onto predict.lm.rrpp}
}
\description{
Print/Summary Function for RRPP
}
\author{
Michael Collyer
}
\keyword{utilities}
