% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRPP.utils.r
\name{print.predict.lm.rrpp}
\alias{print.predict.lm.rrpp}
\title{Print/Summary Function for RRPP}
\usage{
\method{print}{predict.lm.rrpp}(x, PC = FALSE, ...)
}
\arguments{
\item{x}{Object from \code{\link{predict.lm.rrpp}}}

\item{PC}{Logical argument for whether to use predicted values 
rotated to their PCs}

\item{...}{Other arguments passed onto predict.lm.rrpp}
}
\description{
Print/Summary Function for RRPP
}
\author{
Michael Collyer
}
\keyword{utilities}
