% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRPP.utils.r
\name{print.summary.manova.lm.rrpp}
\alias{print.summary.manova.lm.rrpp}
\title{Print/Summary Function for RRPP}
\usage{
\method{print}{summary.manova.lm.rrpp}(x, ...)
}
\arguments{
\item{x}{Object from \code{\link{summary.manova.lm.rrpp}}}

\item{...}{Other arguments passed onto summary.manova.lm.rrpp}
}
\description{
Print/Summary Function for RRPP
}
\author{
Michael Collyer
}
\keyword{utilities}
