% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRPP.utils.r
\name{print.summary.pairwise}
\alias{print.summary.pairwise}
\title{Print/Summary Function for RRPP}
\usage{
\method{print}{summary.pairwise}(x, ...)
}
\arguments{
\item{x}{Object from \code{\link{summary.pairwise}}}

\item{...}{Other arguments passed onto summary.pairwise}
}
\description{
Print/Summary Function for RRPP
}
\author{
Michael Collyer
}
\keyword{utilities}
