% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRPP.utils.r
\name{print.trajectory.analysis}
\alias{print.trajectory.analysis}
\title{Print/Summary Function for RRPP}
\usage{
\method{print}{trajectory.analysis}(x, ...)
}
\arguments{
\item{x}{Object from \code{\link{trajectory.analysis}}}

\item{...}{Other arguments passed onto}
}
\description{
Print/Summary Function for RRPP
}
\author{
Michael Collyer
}
\keyword{utilities}
