% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRPP.utils.r
\name{print.z.model.test}
\alias{print.z.model.test}
\title{Print/Summary Function for RRPP}
\usage{
\method{print}{z.model.test}(x, ...)
}
\arguments{
\item{x}{Object from \code{\link{z.model.test}}}

\item{...}{Other arguments passed onto model.comparison}
}
\description{
Print/Summary Function for RRPP
}
\author{
Michael Collyer
}
\keyword{utilities}
