% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRPP.support.code.r
\docType{data}
\name{Pupfish}
\alias{Pupfish}
\title{Landmarks on pupfish}
\description{
Landmark data from Cyprinodon pecosensis body shapes, with indication of Sex and
Population from which fish were sampled (Marsh or Sinkhole).
}
\details{
These data were previously aligned with GPA.  Centroid size (CS) is also provided.  
See the \link{geomorph} package for details.
}
\references{
Collyer, M.L., D.J. Sekora, and D.C. Adams. 2015. A method for analysis of phenotypic
change for phenotypes described by high-dimensional data. Heredity. 113: doi:10.1038/hdy.2014.75.
}
\author{
Michael Collyer
}
\keyword{datasets}
