% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRPP.data.r
\docType{data}
\name{PupfishMR}
\alias{PupfishMR}
\title{Experimental metabolic rate data from pupfish}
\description{
Simulated metabolic rate data from experimental pupfish
}
\details{
These data are simulated data meant to mimic a subset of the 
data from Collyer & Stockwell (2004).  Variables include experimental treatment 
(Control versus artificially infected), mass in grams, standard length in mm,
and metabolic rate (oxygen consumption, mg O2 per L per hour).  These data do not have
a temporal component like the original data, and do not include aberrant or missing
observations.  These data are suitable data for demonstrating linear regression
and analysis of covariance 
(ANCOVA), used in the book, Permutational Biometry (Collyer & Adams, 2025).
}
\references{
Collyer, M. L., & Stockwell, C. A. (2004). Experimental 
evidence for costs of parasitism for a threatened species, 
White Sands pupfish (Cyprinodon tularosa). Journal of 
Animal Ecology, 73(5), 821-830.

Collyer, M. L., & Adams D. C. (2025). Permutational Biometry: Volume
1, Univariate Data. Iowa State University Digital Press.  NOT YET PUBLISHED.
}
\author{
Michael Collyer
}
