% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared.support.code.r
\name{pval}
\alias{pval}
\title{Obtain P-value from a vector of values}
\usage{
pval(s, target = NULL, greater = TRUE)
}
\arguments{
\item{s}{The sampling distribution vector to use.}

\item{target}{The value to target in the distribution.  (If null, the first value
in the vector is used.).  If the target exists outside the range of s,
a probability of 0 or 1 is certain.}

\item{greater}{Logical value for whether the probability should be "greater than 
or equal to".  Change to greater = FALSE for "less than or equal to".}
}
\description{
A function to find the probability of values greater or lesser than target,
from a vector of values presumably obtained in random permutations.
}
\author{
Michael Collyer
}
\keyword{utilities}
