% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRPP.utils.r
\name{random.effects.lmm.rrpp}
\alias{random.effects.lmm.rrpp}
\title{random.effects for lmm.rrpp model fits}
\usage{
random.effects.lmm.rrpp(x, type, ...)
}
\arguments{
\item{type}{Whether to combine random effects into one matrix or print a list
by variable.}

\item{...}{Other arguments passed.}

\item{object}{Object from \code{\link{lmm.rrpp}}}
}
\description{
Returns random effects from a lmm.rrpp object.  Alias for ranef.lmm.rrpp.
}
\author{
Michael Collyer
}
\keyword{utilities}
