% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRPP.utils.r
\name{ranef.lmm.rrpp}
\alias{ranef.lmm.rrpp}
\title{ranef for lmm.rrpp model fits}
\usage{
ranef.lmm.rrpp(object, type = c("matrix", "list"), ...)
}
\arguments{
\item{object}{Object from \code{\link{lmm.rrpp}}}

\item{type}{Whether to combine random effects into one matrix or print a list
by variable.}

\item{...}{Other arguments passed.}
}
\description{
Returns random effects from a lmm.rrpp object.
}
\author{
Michael Collyer
}
\keyword{utilities}
