% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRPP.data.r
\docType{data}
\name{ReptAbund}
\alias{ReptAbund}
\title{Reptile Abundance Data in Western Australia}
\description{
Reptile abundance data from several sites in Western Australia,
originally collected to examine recovery after a large fire.
}
\details{
The data set is from a study examining reptile abundance from
several sites in Western Australia (Davis and Doherty, 2015). The original 
study examined the recovery of reptile communities in ten remnant
localities following a large fire in 2009. Half of localities were in the burn zone 
and the remaining half were unaffected by the fire. Data were sampled over five years.
These data are suitable data for demonstrating mixed linear models, used in the book, 
Permutational Biometry (Collyer & Adams, 2025).
}
\references{
Davis, R. A. and Doherty, T. S. (2015). Rapid recovery of an urban remnant
reptile community following summer wildfire. PLOS ONE, 10(5):e0127925.

Collyer, M. L., & Adams D. C. (2025). Permutational Biometry: Volume
1, Univariate Data. Iowa State University Digital Press.  NOT YET PUBLISHED.
}
\keyword{datasets}
