% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRPP.data.r
\docType{data}
\name{snakeHS}
\alias{snakeHS}
\title{Snake Head Size and Regional Variation}
\description{
A data set containing variables for snake head size, 
snout-to-vent-length, sex, and geographic information.
}
\details{
These data were collected from prairie rattlesnakes (Crotalus viridis viridis),
from several geographic locations in the upper Midwest of the United States (Smith and 
Collyer, 2008).  Variables include a spatial measure of head size (HS), 
snout-to-vent-length (SVL, cm), sex, and the site and region where 107
snakes were captured.  These 
data are suitable for demonstrating model selection  in the book,
Permutational Biometry (Collyer & Adams, 2025).
}
\references{
Smith, M. T. and Collyer, M. L. (2008). The biology of rattlesnakes, chapter
Regional variation and sexual dimorphism in head form of the prairie
rattlesnake (Crotalus viridid viridis): Comparisons using new analytical techniques
and collection methods, pages 79–90. Loma Linda University Press.

Collyer, M. L., & Adams D. C. (2025). Permutational Biometry: Volume
1, Univariate Data. Iowa State University Digital Press.  NOT YET PUBLISHED.
}
\keyword{datasets}
