% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRPP.utils.r
\name{summary.betaTest}
\alias{summary.betaTest}
\title{Print/Summary Function for RRPP}
\usage{
\method{summary}{betaTest}(object, ...)
}
\arguments{
\item{object}{Object from \code{\link{betaTest}}}

\item{...}{Other arguments passed onto betaTest}
}
\description{
Print/Summary Function for RRPP
}
\author{
Michael Collyer
}
\keyword{utilities}
