% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRPP.utils.r
\name{summary.classify}
\alias{summary.classify}
\title{Print/Summary Function for RRPP}
\usage{
\method{summary}{classify}(object, ...)
}
\arguments{
\item{object}{Object from \code{\link{classify}}}

\item{...}{Other arguments passed onto classify}
}
\description{
Print/Summary Function for RRPP
}
\author{
Michael Collyer
}
\keyword{utilities}
