% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRPP.utils.r
\name{summary.lm.rrpp}
\alias{summary.lm.rrpp}
\title{Print/Summary Function for RRPP}
\usage{
\method{summary}{lm.rrpp}(object, ...)
}
\arguments{
\item{object}{print/summary object (from \code{\link{lm.rrpp}})}

\item{...}{other arguments passed to print/summary}
}
\description{
Print/Summary Function for RRPP
}
\author{
Michael Collyer
}
\keyword{utilities}
