% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRPP.utils.r
\name{summary.lr_test}
\alias{summary.lr_test}
\title{Print/Summary Function for RRPP}
\usage{
\method{summary}{lr_test}(object, ...)
}
\arguments{
\item{object}{Object from \code{\link{lr_test}}}

\item{...}{Other arguments passed onto print.lr_test}
}
\description{
Print/Summary Function for RRPP
}
\author{
Michael Collyer
}
\keyword{utilities}
