% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRPP.utils.r
\name{summary.manova.lm.rrpp}
\alias{summary.manova.lm.rrpp}
\title{Print/Summary Function for RRPP}
\usage{
\method{summary}{manova.lm.rrpp}(object, test = c("Roy", "Pillai",
  "Hotelling-Lawley", "Wilks"), ...)
}
\arguments{
\item{object}{Object from \code{\link{lm.rrpp}}, updated with \code{\link{manova.update}}}

\item{test}{Type of multivariate test statistic to use.}

\item{...}{Other arguments passed onto manova.lm.rrpp}
}
\description{
Print/Summary Function for RRPP
}
\author{
Michael Collyer
}
\keyword{utilities}
