% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRPP.utils.r
\name{summary.measurement.error}
\alias{summary.measurement.error}
\title{Print/Summary Function for RRPP}
\usage{
\method{summary}{measurement.error}(object, ...)
}
\arguments{
\item{object}{Object from \code{\link{measurement.error}}}

\item{...}{Other arguments passed onto measurement.error}
}
\description{
Print/Summary Function for RRPP
}
\author{
Michael Collyer
}
\keyword{utilities}
