% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRPP.utils.r
\name{summary.model.comparison}
\alias{summary.model.comparison}
\title{Print/Summary Function for RRPP}
\usage{
\method{summary}{model.comparison}(object, ...)
}
\arguments{
\item{object}{Object from \code{\link{model.comparison}}}

\item{...}{Other arguments passed onto model.comparison}
}
\description{
Print/Summary Function for RRPP
}
\author{
Michael Collyer
}
\keyword{utilities}
