% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRPP.utils.r
\name{summary.ordinate}
\alias{summary.ordinate}
\title{Print/Summary Function for RRPP}
\usage{
\method{summary}{ordinate}(x, ...)
}
\arguments{
\item{x}{Object from \code{\link{ordinate}}}

\item{...}{Other arguments passed onto print.ordinate}
}
\description{
Print/Summary Function for RRPP
}
\author{
Michael Collyer
}
\keyword{utilities}
