% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRPP.utils.r
\name{summary.pairwise.model.Z}
\alias{summary.pairwise.model.Z}
\title{Print/Summary Function for RRPP}
\usage{
\method{summary}{pairwise.model.Z}(object, ...)
}
\arguments{
\item{object}{print/summary object (from \code{\link{pairwise.model.Z}})}

\item{...}{other arguments passed to print/summary}
}
\description{
Print/Summary Function for RRPP
}
\author{
Dean Adams and Michael Collyer
}
\keyword{utilities}
