% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRPP.utils.r
\name{summary.predict.lm.rrpp}
\alias{summary.predict.lm.rrpp}
\title{Print/Summary Function for RRPP}
\usage{
\method{summary}{predict.lm.rrpp}(object, ...)
}
\arguments{
\item{object}{Object from \code{\link{predict.lm.rrpp}}}

\item{...}{Other arguments passed onto predict.lm.rrpp}
}
\description{
Print/Summary Function for RRPP
}
\author{
Michael Collyer
}
\keyword{utilities}
