% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRPP.utils.r
\name{summary.z.model.test}
\alias{summary.z.model.test}
\title{Print/Summary Function for RRPP}
\usage{
\method{summary}{z.model.test}(object, stat.table = TRUE, names = c("form",
  "model"), ...)
}
\arguments{
\item{stat.table}{Logical value to indicate whether to summarize stats in one table
(default = TRUE) or separate pariwise tables (= FALSE)}

\item{names}{One of "form" or "model.  The former returns formulae for model names; the
latter returns, "m.1", "m.2",...}

\item{...}{Other arguments passed onto model.comparison}

\item{x}{Object from \code{\link{z.model.test}}}
}
\description{
Print/Summary Function for RRPP
}
\author{
Michael Collyer
}
\keyword{utilities}
