% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRPP.utils.r
\name{terms.lm.rrpp}
\alias{terms.lm.rrpp}
\title{Extract the terms from an lm.rrpp object}
\usage{
\method{terms}{lm.rrpp}(x, ...)
}
\arguments{
\item{x}{Object from \code{\link{lm.rrpp}}}

\item{...}{further arguments passed to or from other methods}
}
\description{
\code{terms.lm.rrpp} returns the terms constructed for an \code{lm.rrpp} object.
}
\author{
Michael Collyer
}
\keyword{utilities}
