% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRPP.support.code.r
\name{vec.cor.matrix}
\alias{vec.cor.matrix}
\title{Support function for RRPP}
\usage{
vec.cor.matrix(M)
}
\arguments{
\item{M}{Matrix for vector correlations.}
}
\description{
Calculate vector correlations for a matrix (by rows).  Used for pairwise comparisons.
}
\author{
Michael Collyer
}
\keyword{utilities}
