% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRPP.utils.r
\name{anova.lm.rrpp}
\alias{anova.lm.rrpp}
\title{ANOVA for lm.rrpp model fits}
\usage{
\method{anova}{lm.rrpp}(object, ..., effect.type = c("F", "cohenf", "SS",
  "MS", "Rsq"), error = NULL)
}
\arguments{
\item{object}{Object from \code{\link{lm.rrpp}}}

\item{...}{Other arguments passed to ANOVA}

\item{effect.type}{One of "F", "cohenf", "SS", "MS", "Rsq" to choose from which distribution
of statistics to calculate effect sizes (Z).  See \code{\link{lm.rrpp}}.}

\item{error}{An optional character string to define MS error term for calculation of F values.
See \code{\link{lm.rrpp}} for examples.}
}
\description{
Computes an analysis of variance table using distributions of random
statistics from \code{\link{lm.rrpp}}.
}
\author{
Michael Collyer
}
\keyword{utilities}
