% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRPP.utils.r
\name{fitted.lm.rrpp}
\alias{fitted.lm.rrpp}
\title{Extract fitted values}
\usage{
\method{fitted}{lm.rrpp}(object, weighted = TRUE, ...)
}
\arguments{
\item{object}{plot object (from \code{\link{lm.rrpp}})}

\item{weighted}{A logical argument to return weighted or unweighted fitted values.}

\item{...}{Arguments passed to other functions}
}
\description{
Extract fitted values
}
\examples{
# See examples for lm.rrpp
}
\author{
Michael Collyer
}
\keyword{utilities}
