% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRPP.utils.r
\name{getModels}
\alias{getModels}
\title{Utility Function for RRPP}
\usage{
getModels(fit, attribute = c("terms", "X", "qr", "all"))
}
\arguments{
\item{fit}{Object from \code{\link{lm.rrpp}}}

\item{attribute}{The various attributes that are used to extract RRPP
Model information, including "terms", "X" (design matrices), "qr" (QR
decompositions), or "all".}
}
\description{
A function mostly for internal processing but can be used to obtain terms,
design matrices, or QR decompositions used for each reduced or full
model that is fitted in an \code{\link{lm.rrpp}} fit.
}
\author{
Michael Collyer
}
\keyword{utilities}
