% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manova.update.r
\name{manova.update}
\alias{manova.update}
\title{MANOVA update for lm.rrpp model fits}
\usage{
manova.update(fit, error = NULL, print.progress = TRUE,
  verbose = FALSE)
}
\arguments{
\item{fit}{Linear model fit from \code{\link{lm.rrpp}}}

\item{error}{An optional character string to define R matrices used to calculate invR.H.
(Currently only Residuals can be used and this argument defaults to NULL.  Future versions
will update this argument.)}

\item{print.progress}{A logical value to indicate whether a progress bar should be printed to the screen.
This is helpful for long-running analyses.}

\item{verbose}{A logical value to indicate whether invR.H matrices should be returned (if TRUE) or just their 
eigen values (if FALSE).}
}
\value{
An object of class \code{lm.rrpp} is updated to include class \code{manova.lm.rrpp}, and the object,
$MANOVA, which includes
\item{eiges}{The eigenvalues of invR.H, if verbose = FALSE.}
\item{invR.H}{The inverse of the residuals SSCP times the H SSCP, if verbose = TRUE}
\item{e.rank}{Rank of the error (residuals) covariance matrix.}
}
\description{
Function updates a lm.rrpp fit to add $MANOVA, which like $ANOVA, provides statistics
or matrices typically associated with multivariate analysis of variance (MANOVA).

MANOVA statistics or sums of squares and cross-products (SSCP) matrices
are calculated over the random permutations of a \code{\link{lm.rrpp}} fit.  SSCP matrices are 
computed as the inverse of R time H (invR.H), where R is a SSCP for the residuals or random effects and H is
the difference between SSCP matrices of full and reduced models (see below).  If
the argument, verbose, is TRUE, invR.H matrices will be returned for every model effect and the entire model,
for every permutation.  If verbose = FALSE, eigen-analysis is performed on the invR.H matrices and returned, allowing
MANOVA statistics to be calculated, including Roy's maximum root (eigen value), Pillai trace, Hotelling-Lawley trace,
and Wilks lambda (via \code{\link{summary.manova.lm.rrpp}}).

The manova.update to add $MANOVA to \code{\link{lm.rrpp}} fits requires more computation time than the $ANOVA
statistics that are computed automatically in \code{\link{lm.rrpp}}.  Generally, the same inferential conclusions will
be found with either approach.  Because distributions of MANOVA stats can be generated from the random permutations,
there is no need to approximate F-values, like with parametric MANOVA.  Data dimensionality is not an issue with the 
non-parametric approach used in this function.  By restricting analysis to the real, positive eigen values calculated,
all statistics can be calculated (but Wilks lambda, as a product but not a trace, might be less reliable as variable number approaches
or exceeds the number of observations).

 \subsection{ANOVA vs. MANOVA}{ 
 Two SSCP matrices are calculated for each linear model effect, for every random permutation: R (Residuals or Random effects) and
 H, the difference between SSCPs for "full" and "reduced" models. (Full models contain and reduced models lack
 the effect tested; SSCPs are hypothesized to be the same under a null hypothesis, if there is no effect.  The 
 difference, H, would have a trace of 0 if the null hypothesis were true.)  In RRPP, ANOVA and MANOVA correspond to
 two different ways to calculate statistics from R and H matrices.
 
 ANOVA statistics are those that find the trace of R and H SSCP matrices before calculating subsequent statistics,
 including sums of squares (SS), mean squares (MS), and F-values.  These statistics can be calculated with univariate data
 and provide univariate-like statistics for multivariate data.  These statistics are dispersion measures only (covariances
 among variables do not contribute) and are the same as "distance-based" stats proposed by Goodall (1991) and Anderson (2001).
 MANOVA stats require multivariate data and are implicitly affected by variable covariances.  For MANOVA, the inverse of R times
 H (invR.H) is first calculated for each effect, then eigenanalysis is performed on these matrix products.  Multivariate
 statistics are calculated from the positive, real eigenvalues.  In general, inferential
 conclusions will be similar with either approach, but effect sizes might differ.
 
 ANOVA tables are generated by \code{\link{anova.lm.rrpp}} on lm.rrpp fits and MANOVA tables are generated
 by \code{\link{summary.manova.lm.rrpp}}, after running manova.update on lm.rrpp fits.
 
 Currently, mixed model effects are only possible with $ANOVA statistics, not $MANOVA.
 
}
}
\examples{
   
# Body Shape Analysis (Multivariate) ----------------------------------------------------

data(Pupfish)

# Although not recommended as a practice, this example will use only
# three principal components of body shape for demonstration.  A larger
# number of random permutations should also be used.

Pupfish$shape <- prcomp(Pupfish$coords)$x[, 1:3]

Pupfish$logSize <- log(Pupfish$CS) # better to not have functions in formulas

fit <- lm.rrpp(shape ~ logSize + Sex, SS.type = "I", 
data = Pupfish, print.progress = FALSE, iter = 499) 
summary(fit, formula = FALSE)
anova(fit) # ANOVA table

# MANOVA

fit.m <- manova.update(fit, print.progress = FALSE)
summary(fit.m, test = "Roy")
summary(fit.m, test = "Pillai")

fit.m$MANOVA$eigs$logSize[1:3] # eigenvalues first three iterations
fit.m$MANOVA$eigs$Sex[1:3] # eigenvalues first three iterations

fit.m <- manova.update(fit, verbose = TRUE, print.progress = FALSE)
summary(fit.m, test = "Roy")
summary(fit.m, test = "Pillai")

fit.m$MANOVA$invR.H$logSize[1:3] # invR.H first three iterations
fit.m$MANOVA$invR.H$Sex[1:3] # invR.H first three iterations

# Distributions of test statistics

summ.roy <- summary(fit.m, test = "Roy")
dens <- apply(summ.roy$rand.stats, 2, density)
par(mfcol = c(1, length(dens)))
for(i in 1:length(dens)) {
     plot(dens[[i]], xlab = "Roy max root", ylab = "Density",
     type = "l", main = names(dens)[[i]])
     abline(v = summ.roy$rand.stats[1, i], col = "red")
}
par(mfcol = c(1,1))

}
\references{
Goodall, C.R. 1991. Procrustes methods in the statistical analysis of shape. Journal of the 
   Royal Statistical Society B 53:285-339.

Anderson MJ. 2001. A new method for non-parametric multivariate analysis of variance.
   Austral Ecology 26: 32-46.
}
\author{
Michael Collyer
}
\keyword{analysis}
