% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRPP.utils.r
\name{model.matrix.lm.rrpp}
\alias{model.matrix.lm.rrpp}
\title{Construct Design matrix}
\usage{
\method{model.matrix}{lm.rrpp}(x)
}
\arguments{
\item{x}{Object from \code{\link{lm.rrpp}}}
}
\description{
\code{model.matrix.lm.rrpp} returns the design matrix constructed for an \code{lm.rrpp} object.
}
\author{
Michael Collyer
}
\keyword{utilities}
