% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRPP.data.r
\docType{data}
\name{phytoplankton}
\alias{phytoplankton}
\title{Phytoplankton Biomass and Eutrophication}
\description{
A data set containing variables for phytoplankton biomass,
phosphorous levels, trophic status, and locations (coordinates) of 70 temperate lakes.
}
\details{
The data in this data set are samples from a larger study 
that compared phytoplantkton biomass
from temperate lakes in Alberta, Canada, with different levels of eutrophication
(Loewen et al., 2020).  These 
data are suitable for considering the spatial non-independence of observations in 
generalized least squares (GLS) estimation, for the example in the book,
Permutational Biometry (Collyer & Adams, 2025).
}
\references{
Loewen, C. J. G., Wyatt, F. R., Mortimer, C. A., Vinebrooke, R. D., and
Zurawell, R. W. (2020). Multiscale drivers of phytoplankton communities
in north‐temperate lakes. Ecological Applications, 30(5):e02102.

Collyer, M. L., & Adams D. C. (2025). Permutational Biometry: Volume
1, Univariate Data. Iowa State University Digital Press.  NOT YET PUBLISHED.
}
\keyword{datasets}
