% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRPP.utils.r
\name{plot.measurement.error}
\alias{plot.measurement.error}
\title{Print/Summary Function for RRPP}
\usage{
\method{plot}{measurement.error}(x, add.legend = TRUE, ...)
}
\arguments{
\item{x}{Object from \code{\link{measurement.error}}}

\item{add.legend}{A logical value for whether to add a legend.}

\item{...}{Other arguments passed onto plot}
}
\description{
Print/Summary Function for RRPP
}
\author{
Michael Collyer
}
\keyword{utilities}
