% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRPP.utils.r
\name{print.compare.models}
\alias{print.compare.models}
\title{compare.models}
\usage{
\method{print}{compare.models}(x, confidence = 0.95, method = c("AIC",
  "aRsq", "pcRsq", "LRT", "all"), ...)
}
\arguments{
\item{x}{print/summary object (from \code{\link{lm.rrpp}})}

\item{confidence}{The confidence level for confidence intervals on model selection statistics}

\item{method}{One of AIC, adjusted R-squared (aRsq), first PC R-squared inprovement (pcRsq),
likelihood ratio tests (LRT), or all of these statistics for model comaprisons printed in the summary.}

\item{...}{other arguments passed to print/summary}
}
\description{
Print/Summary Function for RRPP
}
\author{
Michael Collyer
}
\keyword{utilities}
