% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRPP.utils.r
\name{residuals.lm.rrpp}
\alias{residuals.lm.rrpp}
\title{Extract residuals}
\usage{
\method{residuals}{lm.rrpp}(object, weighted = TRUE, ...)
}
\arguments{
\item{object}{plot object (from \code{\link{lm.rrpp}})}

\item{weighted}{A logical argument to return weighted or unweighted residuals.}

\item{...}{Arguments passed to other functions}
}
\description{
Extract residuals
}
\examples{
# See examples for lm.rrpp
}
\author{
Michael Collyer
}
\keyword{utilities}
