% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRPP.support.code.r
\docType{package}
\name{RRPP-package}
\alias{RRPP-package}
\alias{RRPP}
\title{Linear model evaluation with Randomized Residual Permutation Procedures}
\description{
Functions in this package allow one to evaluate linear models with residual randomization.
The name, "RRPP", is an acronym for, "Randomized Residual Permutation Procedure."  Through
the various functions in this package, one can use randomization of residuals to generate empirical probability
distributions for linear model effects, for high-dimensional data or distance matrices.
}
\section{RRPP TOC}{

RRPP-package
}

\author{
Michael Collyer and Dean Adams
}
