% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRPP.utils.r
\name{summary.pairwise}
\alias{summary.pairwise}
\title{Print/Summary Function for RRPP}
\usage{
\method{summary}{pairwise}(object, stat.table = TRUE,
  test.type = c("dist", "VC", "var"), angle.type = c("rad", "deg"),
  confidence = 0.95, show.vectors = FALSE, ...)
}
\arguments{
\item{object}{Object from \code{\link{pairwise}}}

\item{stat.table}{Logical argument for whether results should be returned in one table 
(if TRUE) or separate pairwise tables (if FALSE)}

\item{test.type}{Whether distances or vector correlations between vectors or variances (dispersion of residuals)
should be used in the test.}

\item{angle.type}{If test.type = "VC", whether angle results are expressed in radians or degrees.}

\item{confidence}{Confidence level to use for upper confidence limit; default = 0.95 (alpha = 0.05)}

\item{show.vectors}{Logical value to indicate whether vectors should be printed.}

\item{...}{Other arguments passed onto pairwise}
}
\description{
Print/Summary Function for RRPP
}
\author{
Michael Collyer
}
\keyword{utilities}
